requirements.txt
absl-py==1.4.0
anyio==3.6.2
appdirs==1.4.4
astroid==2.15.1
astunparse==1.6.3
attrs==21.4.0
autograd==1.5
autoray==0.6.3
cachetools==5.3.0
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
circ==0.0.1
cirq==1.1.0
cirq-aqt==1.1.0
cirq-core==1.1.0
cirq-google==1.1.0
cirq-ionq==1.1.0
cirq-pasqal==1.1.0
cirq-rigetti==1.1.0
cirq-web==1.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cma==3.3.0
colorama==0.4.6
contourpy==1.0.7
coverage==7.2.2
cryptography==40.0.2
cycler==0.11.0
deco==0.6.3
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
duet==0.2.7
exceptiongroup==1.1.1
fastdtw==0.3.4
filelock==3.10.7
Fiona==1.9.2
fire==0.5.0
flatbuffers==23.5.9
fonttools==4.39.3
future==0.18.3
gast==0.4.0
geopandas==0.12.2
google-api-core==1.34.0
google-auth==2.17.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.48.2
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.22
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
iso8601==1.1.0
isort==5.12.0
jax==0.4.10
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lark==0.11.3
lazy-object-proxy==1.9.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
ml-dtypes==0.1.0
moody==0.2.2
mplhep==0.3.28
mplhep-data==0.0.3
mpmath==1.3.0
msgpack==1.0.5
munch==2.5.0
mushroom-rl==1.9.1
networkx==2.8.8
nodeenv==1.7.0
ntlm-auth==1.5.0
numpy==1.23.5
numpy-ml==0.1.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
opytimark==1.0.8
opytimizer==3.1.2
packaging==23.0
pandas==1.5.3
pbr==5.11.1
PennyLane==0.29.1
PennyLane-Lightning==0.29.0
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
ply==3.11
pre-commit==3.2.1
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
py==1.11.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
pygame==2.3.0
PyJWT==2.6.0
pylatexenc==2.10
pylint==2.17.1
pyparsing==3.0.9
pyproj==3.5.0
pyquil==3.3.4
pyspnego==0.9.0
pytest==7.2.2
python-constraint==1.4.0
python-dateutil==2.8.2
python-rapidjson==1.10
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qcs-api-client==0.21.3
qibo==0.1.10
qiskit==0.43.0
qiskit-aer==0.12.0
qiskit-ibmq-provider==0.20.2
qiskit-ignis==0.7.1
qiskit-machine-learning==0.6.1
qiskit-terra==0.24.0
requests==2.28.2
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
retry==0.9.2
retrying==1.3.4
retworkx==0.12.1
rfc3339==6.2
rfc3986==1.5.0
rpcq==3.11.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
rustworkx==0.12.1
scikit-learn==1.1.1
scikit-optimize==0.9.0
scipy==1.10.1
semantic-version==2.10.0
sh==2.0.3
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stevedore==5.1.0
sympy==1.11.1
tabulate==0.9.0
tensorflow-macos==2.12.0
tensorboard==2.12.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
torch==2.0.0
tqdm==4.65.0
triple-e @ https://github.com/vbelis/triple_e/archive/master.zip
types-python-dateutil==2.8.19.11
types-retry==0.9.9.3
typing_extensions==4.5.0
uhi==0.3.3
urllib3==1.26.15
virtualenv==20.21.0
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.4
wincertstore==0.2
wrapt==1.14.1
zipp==3.15.0
